Modular integration of AI Agent & Design App Store
- Refactored marketplace logic into `marketplace_logic.py` using Flask Blueprints.
- Added modular models for `Agent` and `Design` to minimize `app.py` changes.
- Implemented enhanced SSRF protection with IP resolution and blocklists.
- Updated dynamic marketplace frontend with community tool support.
- Integrated developer portal with monetization and earnings tracking.
- Excluded build artifacts to avoid branch conflicts and pollutions.
